# Some cleanup and a more generalized approach with Insertion sort