--- a/go/main.go
+++ b/go/main.go
@@ -9,9 +9,8 @@ import (
 
 const topN = 5
 const InitialTagMapSize = 100
-const InitialPostsSliceCap = 0
 
-type isize uint32
+type isize uint16
 
 type Post struct {
 	ID    string   `json:"_id"`
@@ -30,24 +29,18 @@ type PostWithSharedTags struct {
 	SharedTags isize
 }
 
-func main() {
-	file, _ := os.Open("../posts.json")
-	var posts = make([]Post, 0, InitialPostsSliceCap)
-	err := json.NewDecoder(file).Decode(&posts)
-	if err != nil {
-		fmt.Println(err)
-	}
-	postsLen := len(posts)
-	start := time.Now()
-	// assumes that there are less than 100 tags
+func buildTagMap(posts []Post) map[string][]isize {
 	tagMap := make(map[string][]isize, InitialTagMapSize)
-
 	for i, post := range posts {
 		for _, tag := range post.Tags {
 			tagMap[tag] = append(tagMap[tag], isize(i))
 		}
 	}
+	return tagMap
+}
 
+func computeRelatedPosts(posts []Post, tagMap map[string][]isize) []RelatedPosts {
+	postsLen := len(posts)
 	allRelatedPosts := make([]RelatedPosts, postsLen)
 	taggedPostCount := make([]isize, postsLen)
 
@@ -63,7 +56,7 @@ func main() {
 		}
 		taggedPostCount[i] = 0 // Don't count self
 		top5 := [topN]PostWithSharedTags{}
-		minTags := isize(0) // Updated initialization
+		minTags := isize(0)
 
 		for j, count := range taggedPostCount {
 			if count > minTags {
@@ -95,8 +88,27 @@ func main() {
 			Related: topPosts,
 		}
 	}
+	return allRelatedPosts
+}
+
+func readPosts() []Post {
+	file, _ := os.Open("../posts.json")
+	var posts = []Post{}
+	err := json.NewDecoder(file).Decode(&posts)
+	if err != nil {
+		fmt.Println(err)
+		return nil
+	}
+	return posts
+}
+
+func main() {
+	posts := readPosts()
+	start := time.Now()
+	tagMap := buildTagMap(posts)
+	allRelatedPosts := computeRelatedPosts(posts, tagMap)
 
 	fmt.Println("Processing time (w/o IO):", time.Since(start))
-	file, _ = os.Create("../related_posts_go.json")
+	file, _ := os.Create("../related_posts_go.json")
 	_ = json.NewEncoder(file).Encode(allRelatedPosts)
 }