--- a/go/main.go
+++ b/go/main.go
@@ -10,12 +10,17 @@ import (
 const topN = 5
 const InitialTagMapSize = 100
 
+const (
+	INPUT_FILE  = "../posts.json"
+	OUTPUT_FILE = "../related_posts_go.json"
+)
+
 type isize uint16
 
 type Post struct {
-	ID    string   `json:"_id"`
-	Title string   `json:"title"`
-	Tags  []string `json:"tags"`
+	ID    string    `json:"_id"`
+	Title string    `json:"title"`
+	Tags  *[]string `json:"tags"`
 }
 
 type RelatedPosts struct {
@@ -32,7 +37,7 @@ type PostWithSharedTags struct {
 func buildTagMap(posts []Post) map[string][]isize {
 	tagMap := make(map[string][]isize, InitialTagMapSize)
 	for i, post := range posts {
-		for _, tag := range post.Tags {
+		for _, tag := range *post.Tags {
 			tagMap[tag] = append(tagMap[tag], isize(i))
 		}
 	}
@@ -43,27 +48,23 @@ func computeRelatedPosts(posts []Post, tagMap map[string][]isize) []RelatedPosts
 	allRelatedPosts := make([]RelatedPosts, 0, len(posts))
 	taggedPostCount := make([]isize, len(posts))
 
-	for i := range posts {
+	for postIdx := range posts {
 		// Reset the tag count for each post
 		for j := range taggedPostCount {
 			taggedPostCount[j] = 0
 		}
 
 		// Count the number of tags shared between posts
-		for _, tag := range posts[i].Tags {
+		for _, tag := range *posts[postIdx].Tags {
 			for _, otherPostIdx := range tagMap[tag] {
 				taggedPostCount[otherPostIdx]++
 			}
 		}
-		taggedPostCount[i] = 0 // Don't count self
-
 		top5 := [topN]PostWithSharedTags{}
-
+		taggedPostCount[postIdx] = 0 // Don't count self
 		for j, count := range taggedPostCount {
 			if count > top5[0].SharedTags {
 				top5[0] = PostWithSharedTags{Post: isize(j), SharedTags: count}
-
-				// Insertion sort to maintain the order
 				for k := 0; k < topN-1; k++ {
 					if top5[k].SharedTags > top5[k+1].SharedTags {
 						top5[k], top5[k+1] = top5[k+1], top5[k]
@@ -78,8 +79,8 @@ func computeRelatedPosts(posts []Post, tagMap map[string][]isize) []RelatedPosts
 		}
 
 		allRelatedPosts = append(allRelatedPosts, RelatedPosts{
-			ID:      posts[i].ID,
-			Tags:    &posts[i].Tags,
+			ID:      posts[postIdx].ID,
+			Tags:    posts[postIdx].Tags,
 			Related: topPosts,
 		})
 	}
@@ -88,7 +89,7 @@ func computeRelatedPosts(posts []Post, tagMap map[string][]isize) []RelatedPosts
 }
 
 func readPosts() []Post {
-	file, _ := os.Open("../posts.json")
+	file, _ := os.Open(INPUT_FILE)
 	var posts = []Post{}
 	err := json.NewDecoder(file).Decode(&posts)
 	if err != nil {
@@ -105,6 +106,6 @@ func main() {
 	allRelatedPosts := computeRelatedPosts(posts, tagMap)
 
 	fmt.Println("Processing time (w/o IO):", time.Since(start))
-	file, _ := os.Create("../related_posts_go.json")
+	file, _ := os.Create(OUTPUT_FILE)
 	_ = json.NewEncoder(file).Encode(allRelatedPosts)
 }