--- a/julia/related.jl
+++ b/julia/related.jl
@@ -9,6 +9,7 @@ function relatedIO()
     json_string = read("../posts.json", String)
     posts = JSON3.read(json_string, Vector{PostData})    
 
+    related(posts)
     start = now()
     all_related_posts = related(posts)
     println("Processing time (w/o IO): $(now() - start)")