… go/go.sum
… go_con/go.sum