About 5-20% faster, cleaner code and more generalized approach