@DrBlury 2023-10-12
Notice I moved the julia type assertion into the compute time because it's not i/o.