@algunion 2023-10-12
> Notice I moved the julia type assertion into the compute time because it's not i/o.

No, you didn't :)

The Julia time will be now: 2 complete calls of the `related` function + the compilation time of the `related` function (and the other functions called by the `related` function).

I am convinced that you are well intended... but I think it would be fair to do minimal research before deciding to make that kind of change.