@Moelf 2023-10-12
if you're unhappy with Julia's manual warmup, we can revive:
- https://github.com/jinyus/related_post_gen/pull/163


Basically to have native code cached, you want to put the thing into a project. We wanted to avoiding making more directory structures so we opted for a manual call as warmup and cache native code. It does not cache result or value, just a way to trigger JIT compilation.