@DrBlury 2023-10-12
That is very odd. On my machine:
Old:
Benchmark 1: ./related
```
Processing time (w/o IO): 23.913833ms
Processing time (w/o IO): 18.819125ms
Processing time (w/o IO): 18.387583ms
Processing time (w/o IO): 18.44075ms
Processing time (w/o IO): 18.353083ms
Processing time (w/o IO): 19.247333ms
Processing time (w/o IO): 18.802625ms
Processing time (w/o IO): 19.169666ms
Processing time (w/o IO): 18.553125ms
Processing time (w/o IO): 18.385917ms
Processing time (w/o IO): 18.409042ms
Processing time (w/o IO): 18.298292ms
Processing time (w/o IO): 18.476041ms
Processing time (w/o IO): 18.835792ms
Processing time (w/o IO): 18.533083ms
  Time (mean ± σ):      33.5 ms ±   0.6 ms    [User: 31.7 ms, System: 4.5 ms]
  Range (min … max):    32.6 ms …  34.5 ms    10 runs
```

New:
```
Processing time (w/o IO): 19.9455ms
Processing time (w/o IO): 17.655834ms
Processing time (w/o IO): 17.452ms
Processing time (w/o IO): 17.527083ms
Processing time (w/o IO): 17.534792ms
Processing time (w/o IO): 17.525375ms
Processing time (w/o IO): 17.648292ms
Processing time (w/o IO): 17.501708ms
Processing time (w/o IO): 17.549584ms
Processing time (w/o IO): 17.501291ms
Processing time (w/o IO): 17.597291ms
Processing time (w/o IO): 17.441834ms
Processing time (w/o IO): 18.216708ms
  Time (mean ± σ):      39.2 ms ±   0.8 ms    [User: 38.2 ms, System: 3.6 ms]
  Range (min … max):    38.5 ms …  41.3 ms    10 runs
```

Which is about 5% faster, a few colleagues reported up to 18% increases