@jinyus 2023-10-12
Results from the VM (`go version go1.21.3 linux/amd64`):

old: `docker run -e TEST_NAME='go' -it --rm databench`

```
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 24.101841ms
        Processing time (w/o IO): 24.390945ms
        Processing time (w/o IO): 24.185442ms
        Processing time (w/o IO): 24.168141ms
        Processing time (w/o IO): 24.10044ms
        Processing time (w/o IO): 24.05904ms
        Processing time (w/o IO): 24.938753ms
        Processing time (w/o IO): 24.189342ms
        Processing time (w/o IO): 25.076554ms
        Processing time (w/o IO): 24.07584ms
        Processing time (w/o IO): 24.07574ms
        Processing time (w/o IO): 24.290144ms
        Processing time (w/o IO): 24.014639ms
          Time (mean ± σ):      63.0 ms ±   1.1 ms    [User: 58.9 ms, System: 11.0 ms]
          Range (min … max):    61.3 ms …  64.7 ms    10 runs
         
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 371.413354ms
        Processing time (w/o IO): 370.940647ms
        Processing time (w/o IO): 372.894674ms
          Time (mean ± σ):     518.4 ms ±   8.6 ms    [User: 515.8 ms, System: 33.9 ms]
          Range (min … max):   512.3 ms … 524.5 ms    2 runs
         
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 3.276354916s
        Processing time (w/o IO): 3.286848069s
        Processing time (w/o IO): 3.276418201s
          Time (mean ± σ):      3.765 s ±  0.003 s    [User: 3.787 s, System: 0.118 s]
          Range (min … max):    3.762 s …  3.767 s    2 runs
```

new `docker run -e GIT_REPO="https://github.com/DrBlury/related_post_gen.git" -e TEST_NAME='go' -it --rm databench`:

```
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 28.827736ms
        Processing time (w/o IO): 29.510347ms
        Processing time (w/o IO): 29.127541ms
        Processing time (w/o IO): 28.807536ms
        Processing time (w/o IO): 29.409245ms
        Processing time (w/o IO): 28.949638ms
        Processing time (w/o IO): 28.931938ms
        Processing time (w/o IO): 28.957138ms
        Processing time (w/o IO): 28.799236ms
        Processing time (w/o IO): 28.853737ms
        Processing time (w/o IO): 28.965739ms
        Processing time (w/o IO): 28.886338ms
        Processing time (w/o IO): 28.906238ms
          Time (mean ± σ):      69.1 ms ±   0.7 ms    [User: 62.2 ms, System: 13.2 ms]
          Range (min … max):    68.2 ms …  70.5 ms    10 runs
         
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 436.162085ms
        Processing time (w/o IO): 434.575667ms
        Processing time (w/o IO): 432.86844ms
          Time (mean ± σ):     578.4 ms ±   0.4 ms    [User: 575.8 ms, System: 31.7 ms]
          Range (min … max):   578.1 ms … 578.7 ms    2 runs
         
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 3.826478244s
        Processing time (w/o IO): 3.830682611s
        Processing time (w/o IO): 3.832434699s
          Time (mean ± σ):      4.324 s ±  0.019 s    [User: 4.352 s, System: 0.121 s]
          Range (min … max):    4.311 s …  4.338 s    2 runs

```