@DrBlury 2023-10-12
Okay I see the results but in theory it should be faster due to the more efficient mechanism for maintaining the top 5 list (less shifting of elements). And on my machine as well. Are you sure it's not just a VM spasm that is going evil against me?