@jinyus 2023-10-12
Are you using linux? It's very weird indeed. But the results of my machine matches the perf degradation on the VM. 

Did you run using docker? I can write a smaller docker file just for go if you don't want to build the big image.