@DrBlury 2023-10-12
I'm using MacOS (M1) here. I ran using `./run.sh go` so it's ARM.