@DrBlury 2023-10-12
You could add a minimal dockerfile for go that uses a base image which includes ARM in it's supported platforms. That would make it possible to test in docker.