@jinyus 2023-10-12
They are equal on linux arm. The new version is ~1ms faster:

old:

```
Benchmark 1: ./related
Processing time (w/o IO): 38.729717ms
Processing time (w/o IO): 38.882679ms
Processing time (w/o IO): 38.532477ms
Processing time (w/o IO): 38.519756ms
Processing time (w/o IO): 39.1598ms
Processing time (w/o IO): 38.998638ms
Processing time (w/o IO): 38.982198ms
Processing time (w/o IO): 39.485761ms
Processing time (w/o IO): 38.896917ms
Processing time (w/o IO): 39.178919ms
Processing time (w/o IO): 38.907718ms
Processing time (w/o IO): 38.608556ms
Processing time (w/o IO): 38.779797ms

```

new:

```
Benchmark 1: ./related
Processing time (w/o IO): 38.006301ms
Processing time (w/o IO): 37.994581ms
Processing time (w/o IO): 37.846461ms
Processing time (w/o IO): 37.72482ms
Processing time (w/o IO): 37.835541ms
Processing time (w/o IO): 38.165302ms
Processing time (w/o IO): 38.382104ms
Processing time (w/o IO): 37.987501ms
Processing time (w/o IO): 38.336144ms
Processing time (w/o IO): 38.221782ms
Processing time (w/o IO): 37.938342ms
Processing time (w/o IO): 37.949861ms
Processing time (w/o IO): 37.693219ms
```